+++ packages/universal-modal/src/components/content/index.ts
@@ -1,0 +1,8 @@
+export { ContentResponsive } from './responsive/Component.responsive';
+export type { ContentResponsiveProps } from './responsive/Component.responsive';
+
+export { ContentDesktop } from './desktop/Component.desktop';
+export type { ContentDesktopProps } from './desktop/Component.desktop';
+
+export { ContentMobile } from './mobile/Component.mobile';
+export type { ContentMobileProps } from './mobile/Component.mobile';