+++ packages/universal-modal/src/components/header/responsive/Component.responsive.tsx
@@ -1,0 +1,18 @@
+import React, { FC } from 'react';
+
+import { useIsDesktop } from '@alfalab/core-components-mq';
+
+import { HeaderDesktop, HeaderDesktopProps } from '../desktop/Component.desktop';
+import { HeaderMobile, HeaderMobileProps } from '../mobile/Component.mobile';
+
+export type HeaderResponsiveProps = HeaderDesktopProps | HeaderMobileProps;
+
+export const HeaderResponsive: FC<HeaderResponsiveProps> = ({ ...restProps }) => {
+    const isDesktop = useIsDesktop();
+
+    if (isDesktop) {
+        return <HeaderDesktop {...restProps} />;
+    }
+
+    return <HeaderMobile {...restProps} />;
+};