+++ packages/universal-modal/src/desktop/index.ts
@@ -1,0 +1,14 @@
+export { UniversalModalDesktop, UniversalModalDesktopComponent } from './Component.desktop';
+export type { UniversalModalDesktopProps } from './types/props';
+
+export { HeaderDesktop } from '../components/header';
+export type { HeaderDesktopProps } from '../components/header';
+
+export { FooterDesktop } from '../components/footer';
+export type { FooterDesktopProps } from '../components/footer';
+
+export { ContentDesktop } from '../components/content';
+export type { ContentDesktopProps } from '../components/content';
+
+export { CrossButtonDesktop } from './components/buttons/cross-button';
+export { ArrowButtonDesktop } from './components/buttons/arrow-button';