+++ packages/universal-modal/src/index.ts
@@ -1,0 +1,11 @@
+export { UniversalModalResponsive, UniversalModal } from './Component.responsive';
+export type { UniversalModalResponsiveProps } from './typings';
+
+export { ContentResponsive } from './components/content';
+export type { ContentResponsiveProps } from './components/content';
+
+export { HeaderResponsive } from './components/header';
+export type { HeaderResponsiveProps } from './components/header';
+
+export { FooterResponsive } from './components/footer';
+export type { FooterResponsiveProps } from './components/footer';