+++ packages/universal-modal/src/mobile/index.ts
@@ -1,0 +1,14 @@
+export { UniversalModalMobile, UniversalModalMobileComponent } from './Component.mobile';
+export type { UniversalModalMobileProps } from './types/props';
+
+export { HeaderMobile } from '../components/header';
+export type { HeaderMobileProps } from '../components/header';
+
+export { FooterMobile } from '../components/footer';
+export type { FooterMobileProps } from '../components/footer';
+
+export { ContentMobile } from '../components/content';
+export type { ContentMobileProps } from '../components/content';
+
+export { CrossButtonMobile } from './components/buttons/cross-button';
+export { ArrowButtonMobile } from './components/buttons/arrow-button';