+++ packages/pattern-lock/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @alfalab/core-components-pattern-lock
 
+## 2.5.10
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@11.11.6
+
+## 2.5.9
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.16.0
+    -   button@11.11.5
+
+## 2.5.8
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.15.0
+    -   button@11.11.4
+
+## 2.5.7
+
+### Patch Changes
+
+<sup><time>09.01.2025</time></sup>
+
+### [#1461](https://github.com/core-ds/core-components/pull/1461)
+
+-   Обновление зависимостей
+
+-   Обновлены зависимости
+    -   button@11.11.3
+
 ## 2.5.6
 
 ### Patch Changes