+++ packages/typography/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Change Log
 
+## 4.12.0
+
+### Minor Changes
+
+<sup><time>21.02.2025</time></sup>
+
+### [#1589](https://github.com/core-ds/core-components/pull/1589)
+
+-   Добавлен static-positive цвет для Typography
+
+## 4.11.4
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   skeleton@5.3.0
+
+## 4.11.3
+
+### Patch Changes
+
+<sup><time>09.01.2025</time></sup>
+
+### [#1461](https://github.com/core-ds/core-components/pull/1461)
+
+-   Обновление зависимостей
+
+-   Обновлены зависимости
+    -   skeleton@5.2.3
+
 ## 4.11.2
 
 ### Patch Changes